Extract routes into methods so they can assembled by projects downstream
