+++ core/src/main/scala/org/genivi/sota/core/rvi/RviClient.scala
@@ -6,11 +6,11 @@ package org.genivi.sota.core.rvi
 
 import com.github.nscala_time.time.Imports.DateTime
 import io.circe._
+import org.genivi.sota.core.resolver.ConnectivityClient
+
 import scala.concurrent.ExecutionContext
 import scala.concurrent.Future
 import scala.util.Random
-
-import org.genivi.sota.core.ConnectivityClient
 
 /**
  * Concrete implementation for sending a message to RVI in JSON-RPC format.
