+++ core/src/test/scala/org/genivi/sota/core/FakeExternalResolver.scala
@@ -4,6 +4,7 @@ import akka.actor.ActorSystem
 import akka.http.scaladsl.model.{HttpResponse, Uri}
 import akka.stream.ActorMaterializer
 import io.circe.Json
+import org.genivi.sota.core.resolver.DefaultExternalResolverClient
 import org.genivi.sota.data.{PackageId, Vehicle}
 
 import scala.concurrent.Future
