+++ core/src/test/scala/org/genivi/sota/core/PackageResourceWordSpec.scala
@@ -20,6 +20,7 @@ import org.scalatest.{Matchers, WordSpec}
 import scala.concurrent.Await
 import slick.driver.MySQLDriver.api._
 import DataPackage._
+import org.genivi.sota.core.resolver.DefaultExternalResolverClient
 import org.genivi.sota.data.PackageId
 
 /**
